pyproject.toml
[project]
name = "can-i-stack-it"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "George Alton", email = "georgealton@gmail.com" }
]
dependencies = [
    "boto3>=1.34.116",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "punq>=0.7.0",
    "aws-sam-translator>=1.89.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
can-i-stack-it = "can_i_stack_it.entrypoints.cli:cli"

[tool.rye]
managed = true
dev-dependencies = [
    "boto3-stubs[iam,cloudformation]>=1.34.117",
    "pytest>=8.2.1",
    "nox>=2024.4.15",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/can_i_stack_it"]
